+++ src/app/[lang]/loading.tsx
@@ -1,4 +1,6 @@
+import { Loader } from 'lucide-react';
+
 export default function Loading() {
-    // You can add any UI inside Loading, including a Skeleton.
-    return <h1>Loading...</h1>;
+  // You can add any UI inside Loading, including a Skeleton.
+  return <Loader />;
 }
